Split part of  `needless_borrow` into `needless_borrows_for_generic_args`
